fix(test): declare(strict_types=1)による型エラーを修正
RectorのDeclareStrictTypesTestsRector（PHPUnitSetList::PHPUNIT_CODE_QUALITYに含まれる）
により追加されたdeclare(strict_types=1)によって顕在化した型エラーを修正しました。

主な修正内容:
- DECIMAL型プロパティへのint → string変換（setPrice02, setTaxRate, setStock等）
- boolean型プロパティへのint → bool変換（setOptionProductTaxRule）
- number_format()への型キャスト追加
- sha1()への型キャスト追加
- createProduct()の引数の型修正（'abstract' → true）

+++ tests/Eccube/Tests/Web/Admin/IndexControllerTest.php
@@ -108,7 +108,7 @@ final class IndexControllerTest extends AbstractAdminWebTestCase
         $this->assertTrue($this->client->getResponse()->isSuccessful());
 
         preg_match('/^￥([0-9,]+) \/ ([0-9]+)/u', trim($crawler->filter('#chart-statistics > div.card-body > div.row:nth-child(1) > div:nth-child(2) > div')->text()), $match);
-        $this->expected = number_format($todaysSales);
+        $this->expected = number_format((float) $todaysSales);
         $this->actual = $match[1];
         $this->verify('本日の売上');
 
@@ -117,7 +117,7 @@ final class IndexControllerTest extends AbstractAdminWebTestCase
         $this->verify('本日の売上件数');
 
         preg_match('/^￥([0-9,]+) \/ ([0-9]+)/u', trim($crawler->filter('#chart-statistics > div.card-body > div.row:nth-child(1) > div:nth-child(3) > div')->text()), $match);
-        $this->expected = number_format($yesterdaysSales);
+        $this->expected = number_format((float) $yesterdaysSales);
         $this->actual = $match[1];
         $this->verify('昨日の売上');
 
@@ -126,7 +126,7 @@ final class IndexControllerTest extends AbstractAdminWebTestCase
         $this->verify('昨日の売上件数');
 
         preg_match('/^￥([0-9,]+) \/ ([0-9]+)/u', trim($crawler->filter('#chart-statistics > div.card-body > div.row:nth-child(1) > div:nth-child(1) > div')->text()), $match);
-        $this->expected = number_format((new \DateTime('today'))->format('m') === (new \DateTime('yesterday'))->format('m') ? bcadd($todaysSales, $yesterdaysSales, 2) : $todaysSales);
+        $this->expected = number_format((float) ((new \DateTime('today'))->format('m') === (new \DateTime('yesterday'))->format('m') ? bcadd($todaysSales, $yesterdaysSales, 2) : $todaysSales));
         $this->actual = $match[1];
         $this->verify('今月の売上');
